requirements.txt
anyio==4.9.0
asgiref==3.8.1
certifi==2025.4.26
Django==5.1.2
django-jazzmin==3.0.1
django-redis==5.4.0
et_xmlfile==2.0.0
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
numpy==2.2.0
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-telegram-bot==22.1
pytz==2024.2
redis==6.1.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
whitenoise==6.9.0
